property_manager_api/web/api/work_orders/schema.py
from pydantic import BaseModel


class MaintenaceRequestText(BaseModel):
    """Simple maintenance request message."""
    message: str

class MaintenaceRequest(BaseModel):
    """Simple maintenance request model."""
    data: dict